+++ src/shpg.f90
@@ -173,7 +173,7 @@ integer :: nloc1
 integer :: nloc2
 integer :: nlon
 integer :: nte
-real :: wshp
+real(wp) :: wshp
 real work(*)
 dimension wshp(*), iwshp(*)
 !
@@ -234,8 +234,8 @@ subroutine shpgi_lower_routine(nlat, nlon, isym, mtrunc, idp, ierror, &
   pe, po, ze, zo, ipse, jzse, ipso, jzso, &
   cp, wx, thet, gwts, xx, z, a, b, ped, pod, u)
 
-real :: dfn
-real :: dmax
+real(wp) :: dfn
+real(wp) :: dmax
 integer :: i
 integer :: idp
 integer :: ierr
@@ -271,14 +271,14 @@ integer :: nshe
 integer :: nsho
 integer :: nte
 integer :: nto
-real :: pe
-real :: po
-real :: sum1
-real :: toe
-real :: tusl
-real :: ze
-real :: zo
-real :: zort
+real(wp) :: pe
+real(wp) :: po
+real(wp) :: sum1
+real(wp) :: toe
+real(wp) :: tusl
+real(wp) :: ze
+real(wp) :: zo
+real(wp) :: zort
 !
 real summation, eps, a1, b1, c1, work
 parameter (eps=epsilon(1.0_wp))
@@ -768,11 +768,11 @@ integer :: nloc1
 integer :: nloc2
 integer :: nlon
 integer :: nte
-real :: sn
-real :: work
-real :: wshp
-real :: x
-real :: y
+real(wp) :: sn
+real(wp) :: work
+real(wp) :: wshp
+real(wp) :: x
+real(wp) :: y
 !
 dimension wshp(*), iwshp(*), work(*), x(idxy, nlon), y(idxy, nlon)
 !
@@ -866,16 +866,16 @@ integer :: nshe
 integer :: nsho
 integer :: nte
 integer :: nto
-real :: pe
-real :: po
-real :: sx
-real :: sy
-real :: xe
-real :: xo
-real :: ye
-real :: yo
-real :: ze
-real :: zo
+real(wp) :: pe
+real(wp) :: po
+real(wp) :: sx
+real(wp) :: sy
+real(wp) :: xe
+real(wp) :: xo
+real(wp) :: ye
+real(wp) :: yo
+real(wp) :: ze
+real(wp) :: zo
 !
 dimension sx(idxy, nlon), sy(idxy, nlon), nshe(2), nsho(2), &
   pe(idp, idp, 2), po(idp, idp, 2), ze(idp, idp, 2), zo(idp, idp, 2), &
@@ -1029,9 +1029,9 @@ return
 end subroutine mxm
 subroutine smxm(lr, lc, ld, a, mc, md, b, nd, c)
 
-real :: a
-real :: b
-real :: c
+real(wp) :: a
+real(wp) :: b
+real(wp) :: c
 integer :: i
 integer :: j
 integer :: k
@@ -1055,8 +1055,8 @@ end subroutine smxm
 
 subroutine mxmx(lr, lc, ld, a, mc, md, b, x, y)
 
-real :: a
-real :: b
+real(wp) :: a
+real(wp) :: b
 integer :: i
 integer :: j
 integer :: k
@@ -1065,10 +1065,10 @@ integer :: ld
 integer :: lr
 integer :: mc
 integer :: md
-real :: sum1
-real :: sum2
-real :: x
-real :: y
+real(wp) :: sum1
+real(wp) :: sum2
+real(wp) :: x
+real(wp) :: y
 dimension a(ld, *), b(md, *), x(ld, 2), y(ld, 2)
 do k=1, lr
 y(k, 1) = 0.
@@ -1124,8 +1124,8 @@ return
 end subroutine dmxmx
 subroutine tmxmx(lag, lr, lc, ld, a, mc, md, b, x, y, is, js)
 
-real :: a
-real :: b
+real(wp) :: a
+real(wp) :: b
 integer :: i
 integer :: is
 integer :: j
@@ -1138,10 +1138,10 @@ integer :: ld
 integer :: lr
 integer :: mc
 integer :: md
-real :: sum1
-real :: sum2
-real :: x
-real :: y
+real(wp) :: sum1
+real(wp) :: sum2
+real(wp) :: x
+real(wp) :: y
 dimension a(ld, *), b(md, *), x(ld, 2), y(ld, 2), &
               is(*), js(*)
 !
@@ -1535,7 +1535,7 @@ end subroutine dlftg
 !
 subroutine compute_gaussian_latitudes_and_weights(nlat, theta, wts, w, lwork, ierror)
 
-real :: eps
+real(wp) :: eps
 integer :: i
 integer :: idx
 integer :: ierror
@@ -1546,7 +1546,7 @@ integer :: nhalf
 integer :: nix
 integer :: nlat
 integer :: ns2
-real :: sgnd
+real(wp) :: sgnd
 !
 !                             April 2002
 !
